+++ 05-SceneManager/Koopas.h
@@ -76,10 +76,10 @@ public:
 	CKoopas(float x, float y,float type);
 	int GetLevel() { return level; };
 	void SetLevel(int level) { this->level = level; };
-	int GetType() { return type; };
-	void SetType(int type) { this->type = type; };
-	int GetKoopasType() { return koopa_type; };
-	void SetKoopasType(int koopa_type) { this->koopa_type = koopa_type; };
+	float GetType() { return type; };
+	void SetType(float type) { this->type = type; };
+	float GetKoopasType() { return koopa_type; };
+	void SetKoopasType(float koopa_type) { this->koopa_type = koopa_type; };
 	virtual void SetState(int state);
 	Timer* timeStartJump = new Timer(2000);
 	Timer* timeReborn = new Timer(5000);
